Animation in separate class
Very cool, though the dropdowns aren't really user friendly and you might want to use a loading animation.
Have you tried putting the animation in a separate class and adding it to the section only when a new dropdown is selected. You should, of course, also delete it first (so the transition has to start again). I would also add a transition in the separate class, so when you remove the class it has no transition time but when you add it, it does. 

Hope this makes sense and hope it works :) Also: please add the music, that would make it complete!